requirements.txt
Django==4.2.1
djangorestframework==3.14.0
geopy==2.3.0
psycopg2==2.9.6
